move crossing out to page
